# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Stellar wallet backend — auth, encrypted credential backup, and account recovery.",
        "title": "Latch API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/multisig/accounts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "multisig-accounts"
                ],
                "summary": "List the session user's deployed multisig accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/multisig/accounts/deploy": {
            "post": {
                "description": "Deploys the multisig smart account for the given threshold, signer set, and salt via the factory contract. Idempotent: returns alreadyDeployed=true if it's already live.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "multisig-accounts"
                ],
                "summary": "Deploy a multisig account on-chain",
                "parameters": [
                    {
                        "description": "Threshold, signer set, and account salt",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.deployAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/multisig/accounts/draft": {
            "post": {
                "description": "Computes the deterministic smart account address and factory deploy params for a given threshold and signer set, without persisting anything or deploying on-chain.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "multisig-accounts"
                ],
                "summary": "Predict a multisig account address from signers",
                "parameters": [
                    {
                        "description": "Threshold and signer set",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.draftAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/multisig/accounts/register": {
            "post": {
                "description": "Persists the multisig account and its members for the session user after on-chain deploy, so it appears in List.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "multisig-accounts"
                ],
                "summary": "Register a deployed multisig account for the session user",
                "parameters": [
                    {
                        "description": "Deployed account address, threshold, salt, and members",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.registerAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler_webapp.deployAccountRequest": {
            "type": "object",
            "required": [
                "accountSaltHex",
                "signers",
                "threshold"
            ],
            "properties": {
                "accountSaltHex": {
                    "type": "string"
                },
                "signers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler_webapp.multisigSignerInitRequest"
                    }
                },
                "threshold": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handler_webapp.draftAccountRequest": {
            "type": "object",
            "required": [
                "signers",
                "threshold"
            ],
            "properties": {
                "accountSaltHex": {
                    "type": "string"
                },
                "signers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler_webapp.multisigSignerInitRequest"
                    }
                },
                "threshold": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handler_webapp.multisigSignerInitRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "gAddress": {
                    "type": "string"
                },
                "keyDataHex": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_webapp.registerAccountRequest": {
            "type": "object",
            "required": [
                "accountSaltHex",
                "members",
                "smartAccountAddress",
                "threshold"
            ],
            "properties": {
                "accountSaltHex": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler_webapp.registerMemberRequest"
                    }
                },
                "smartAccountAddress": {
                    "type": "string"
                },
                "threshold": {
                    "type": "integer"
                }
            }
        },
        "internal_handler_webapp.registerMemberRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "credentialId": {
                    "type": "string"
                },
                "gAddress": {
                    "type": "string"
                },
                "keyDataHex": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_webapp.webappErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "internal_error"
                },
                "error": {
                    "type": "string",
                    "example": "internal error"
                },
                "message": {
                    "type": "string",
                    "example": "internal error"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RecoveryAuth": {
            "description": "Enter: Bearer \u003crecovery_token\u003e  (obtained from POST /v1/recovery/verify)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}